Preview item accounting

+++ Gladys/Models/Types/ArchivedDropItemType+Preview.swift
@@ -32,7 +32,7 @@ extension ArchivedDropItemType {
         let needsCleanup: Bool
         let parentUuid: UUID
         let uuid: UUID
-
+        
         init(typeItem: ArchivedDropItemType) {
 
             parentUuid = typeItem.parentUuid
@@ -53,10 +53,10 @@ extension ArchivedDropItemType {
             } else {
                 previewItemURL = blobPath
             }
-
+            
             previewItemTitle = typeItem.oneTitle
         }
-
+        
         deinit {
             if needsCleanup, let previewItemURL = previewItemURL {
                 let fm = FileManager.default
@@ -75,6 +75,8 @@ extension ArchivedDropItemType {
 		let parentUuid: UUID
 		let uuid: UUID
 
+        private static var previewUrls = [URL: Int]()
+
 		init(typeItem: ArchivedDropItemType) {
 
 			parentUuid = typeItem.parentUuid
@@ -86,19 +88,25 @@ extension ArchivedDropItemType {
 			needsCleanup = blobPath != tempPath
 
 			if needsCleanup {
-				let fm = FileManager.default
-				if !fm.fileExists(atPath: tempPath.path) {
-					if tempPath.pathExtension == "webloc", let url = typeItem.encodedUrl { // only happens on macOS, iOS uses another view for previewing
-						try? PropertyListSerialization.data(fromPropertyList: ["URL": url.absoluteString], format: .binary, options: 0).write(to: tempPath)
-						log("Created temporary webloc for preview: \(tempPath.path)")
-					} else if let data = typeItem.dataForDropping {
-						try? data.write(to: tempPath)
-						log("Created temporary file for preview: \(tempPath.path)")
-					} else {
-						try? fm.linkItem(at: blobPath, to: tempPath)
-						log("Linked temporary file for preview: \(tempPath.path)")
-					}
-				}
+                
+                let currentCount = PreviewItem.previewUrls[tempPath] ?? 0
+                PreviewItem.previewUrls[tempPath] = currentCount + 1
+
+                if currentCount == 0 {
+                    let fm = FileManager.default
+                    if !fm.fileExists(atPath: tempPath.path) {
+                        if tempPath.pathExtension == "webloc", let url = typeItem.encodedUrl { // only happens on macOS, iOS uses another view for previewing
+                            try? PropertyListSerialization.data(fromPropertyList: ["URL": url.absoluteString], format: .binary, options: 0).write(to: tempPath)
+                            log("Created temporary webloc for preview: \(tempPath.path)")
+                        } else if let data = typeItem.dataForDropping {
+                            try? data.write(to: tempPath)
+                            log("Created temporary file for preview: \(tempPath.path)")
+                        } else {
+                            try? fm.linkItem(at: blobPath, to: tempPath)
+                            log("Linked temporary file for preview: \(tempPath.path)")
+                        }
+                    }
+                }
 				previewItemURL = tempPath
 			} else {
 				previewItemURL = blobPath
@@ -109,11 +117,17 @@ extension ArchivedDropItemType {
 
 		deinit {
 			if needsCleanup, let previewItemURL = previewItemURL {
-				let fm = FileManager.default
-				if fm.fileExists(atPath: previewItemURL.path) {
-					try? fm.removeItem(at: previewItemURL)
-					log("Removed temporary preview at \(previewItemURL.path)")
-				}
+                let currentCount = PreviewItem.previewUrls[previewItemURL] ?? 0
+                if currentCount == 1 {
+                    PreviewItem.previewUrls[previewItemURL] = nil
+                    let fm = FileManager.default
+                    if fm.fileExists(atPath: previewItemURL.path) {
+                        try? fm.removeItem(at: previewItemURL)
+                        log("Removed temporary preview at \(previewItemURL.path)")
+                    }
+                } else {
+                    PreviewItem.previewUrls[previewItemURL] = currentCount - 1
+                }
 			}
 		}
 	}
